Moved GUID and UUID to core
