+++ formats/src/test/java/io/parsingdata/metal/FormatTest.java
@@ -16,40 +16,30 @@
 
 package io.parsingdata.metal;
 
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import static io.parsingdata.metal.Shorthand.def;
-import static io.parsingdata.metal.Shorthand.eq;
-import static io.parsingdata.metal.format.UID.guid;
-import static io.parsingdata.metal.format.UID.uuid;
 import static io.parsingdata.metal.util.EncodingFactory.enc;
-import static io.parsingdata.metal.util.EncodingFactory.le;
 import static io.parsingdata.metal.util.EnvironmentFactory.stream;
 
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
 
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.ExpectedException;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 
 import io.parsingdata.metal.format.JPEG;
 import io.parsingdata.metal.format.PNG;
-import io.parsingdata.metal.format.UID;
 import io.parsingdata.metal.format.ZIP;
-import io.parsingdata.metal.token.Token;
 
+@RunWith(JUnit4.class)
 public class FormatTest {
 
     private static final String PNGFILE = "/test.png";
     private static final String ZIPFILE1 = "/singlefile-zip30-ubuntu.zip";
     private static final String ZIPFILE2 = "/multifile-zip30-ubuntu.zip";
     private static final String JPEGFILE = "/test.jpg";
-
-    @Rule
-    public ExpectedException _exception = ExpectedException.none();
 
     @Test
     public void parsePNG() throws IOException, URISyntaxException {
@@ -69,25 +59,6 @@ public class FormatTest {
     @Test
     public void parseJPEG() throws IOException, URISyntaxException {
         assertTrue(JPEG.FORMAT.parse(stream(toURI(JPEGFILE)), enc()).succeeded);
-    }
-
-    @Test
-    public void parseGUID() throws IOException, URISyntaxException {
-        final Token guid = def("guid", 16, eq(guid("00c27766-f623-4200-9d64-115e9bfd4a08")));
-        assertTrue("be", guid.parse(stream(0x00, 0xc2, 0x77, 0x66, 0xf6, 0x23, 0x42, 0x00, 0x9d, 0x64, 0x11, 0x5e, 0x9b, 0xfd, 0x4a, 0x08), enc()).succeeded);
-        assertTrue("le", guid.parse(stream(0x66, 0x77, 0xc2, 0x00, 0x23, 0xf6, 0x00, 0x42, 0x9d, 0x64, 0x11, 0x5e, 0x9b, 0xfd, 0x4a, 0x08), le()).succeeded);
-
-        _exception.expect(IllegalArgumentException.class);
-        _exception.expectMessage("Invalid GUID string: test");
-        guid("test");
-        new UID(); // test coverage check
-    }
-
-    @Test
-    public void parseUUID() throws IOException, URISyntaxException {
-        final Token uuid = def("uuid", 16, eq(uuid("00c27766-f623-4200-9d64-115e9bfd4a08")));
-        assertTrue(uuid.parse(stream(0x00, 0xc2, 0x77, 0x66, 0xf6, 0x23, 0x42, 0x00, 0x9d, 0x64, 0x11, 0x5e, 0x9b, 0xfd, 0x4a, 0x08), enc()).succeeded);
-        assertFalse(uuid.parse(stream(0x66, 0x77, 0xc2, 0x00, 0x23, 0xf6, 0x00, 0x42, 0x9d, 0x64, 0x11, 0x5e, 0x9b, 0xfd, 0x4a, 0x08), enc()).succeeded);
     }
 
     private URI toURI(final String resource) throws URISyntaxException {
